+++ test/optimizer/level-2/optimize-test.js
@@ -55,6 +55,10 @@ vows.describe('level 2 optimizer')
       'calc() function as animation duration': [
         '.block{animation:calc(0.3s * 2) linear 500ms backwards name}',
         '.block{animation:calc(0.3s * 2) linear 0.5s backwards name}'
+      ],
+      'default value of transition-delay is default': [
+        '.block-1{transition:opacity .3s linear,height 0s ease .3s}.block-2{transition:opacity .3s linear,height 1s ease .3s}',
+        '.block-1{transition:opacity .3s linear,height 0s .3s}.block-2{transition:opacity .3s linear,height 1s .3s}'
       ]
     }, { level: 2 })
   )